프론트팀 코드 리뷰용 브랜치 머지

See merge request sw_track/class_08/web_project_ii/team08/its-yours2!277

+++ backend/src/controllers/CommentController.js
@@ -1,4 +1,6 @@
 const CommentService = require('../services/CommentService');
+const { BadRequestError, UnauthorizedError, NotFoundError } = require('../config/CustomError');
+const ObjectId = require('mongodb').ObjectId;
 
 /**
  * 커뮤니티 댓글 작성 service
@@ -13,6 +15,12 @@ const createComment = async (req, res, next) => {
 		const profilePic = req.user.profilePic;
 		const { content } = req.body;
 
+		if (!nickName) {
+			throw new BadRequestError('로그인 후 이용해주세요.');
+		}
+		if (!content) {
+			throw new BadRequestError('필수 내용을 입력 후 등록해주세요.');
+		}
 		const newComment = await CommentService.createComment(postId, content, nickName, profilePic);
 
 		res.status(200).json(newComment);
@@ -30,7 +38,6 @@ const createComment = async (req, res, next) => {
 const getComment = async (req, res, next) => {
 	try {
 		const { postId } = req.params;
-		console.log(req.params);
 		const getComment = await CommentService.getComment(postId);
 
 		res.status(200).json(getComment);
@@ -47,13 +54,21 @@ const getComment = async (req, res, next) => {
  */
 const updateComment = async (req, res, next) => {
 	try {
-		const { commentId } = req.params;
+		const writer = req.user.nickName;
+		const { commentId, postId } = req.params;
 		const { content } = req.body;
 
-		const updateResult = await CommentService.updateComment(commentId, content);
+		const commentInfo = await CommentService.getOneComment(commentId, postId);
 
-		if (!updateComment) {
-			return res.status(400).json({ message: '댓글을 찾을 수 없습니다.' });
+		if (!writer) {
+			throw new BadRequestError('로그인 후 이용해주세요.');
+		}
+		if (writer !== commentInfo.nickName) {
+			throw new UnauthorizedError('작성자만 자신의 댓글을 수정할 수 있습니다.');
+		}
+		const updateResult = await CommentService.updateComment(commentId, content);
+		if (!updateResult) {
+			throw new BadRequestError('댓글을 찾을 수 없습니다.');
 		}
 		res.status(200).json(updateResult);
 	} catch (err) {
@@ -69,12 +84,28 @@ const updateComment = async (req, res, next) => {
  */
 const deleteComment = async (req, res) => {
 	try {
-		const { commentId } = req.params;
+		const { commentId, postId } = req.params;
+		const writer = req.user.nickName;
+		const isAdmin = req.user.isAdmin;
+
+		const commentInfo = await CommentService.getOneComment(commentId, postId);
+
+		if (!commentInfo) {
+			throw new NotFoundError('댓글이 없습니다.');
+		}
+		if (!writer) {
+			throw new BadRequestError('로그인 후 이용해주세요.');
+		}
+		if (!isAdmin) {
+			if (writer !== commentInfo.nickName) {
+				throw new UnauthorizedError('작성자만 자신의 댓글을 삭제할 수 있습니다.');
+			}
+		}
 
 		const deleteComment = await CommentService.deleteComment(commentId);
 
 		if (!deleteComment) {
-			return res.status(400).json({ message: '댓글을 찾을 수 없습니다' });
+			throw new BadRequestError('댓글을 찾을 수 없습니다.');
 		}
 		res.status(200).json({ message: '댓글이 성공적으로 삭제되었습니다.', deleteComment });
 	} catch (err) {
